--- pkg/gofr/datasource/file/ftp/ftp_test.go
@@ -1,1 +1,0 @@
-package ftp
